+++ pymc_experimental/bayesian_estimator_linearmodel.py
@@ -1,6 +1,4 @@
-import json
 from abc import abstractmethod
-from pathlib import Path
 from typing import Any, Dict, Union
 
 import arviz as az
@@ -60,7 +58,6 @@ class BayesianEstimator(ModelBuilder):
 
     def __init__(
         self,
-        data: Union[np.ndarray, pd.DataFrame, pd.Series] = None,
         model_config: Dict = None,
         sampler_config: Dict = None,
     ):
@@ -330,44 +327,6 @@ class BayesianEstimator(ModelBuilder):
         )
 
         return posterior_predictive_samples
-
-    @classmethod
-    def load(cls, fname: str):
-        """
-        Creates a BayesianEstimator instance from a file,
-        Loads inference data for the model.
-
-        Parameters
-        ----------
-        fname : string
-            This denotes the name with path from where idata should be loaded from.
-
-        Returns
-        -------
-        Returns an instance of BayesianEstimator.
-
-        Raises
-        ------
-        ValueError
-            If the inference data that is loaded doesn't match with the model.
-        """
-
-        filepath = Path(str(fname))
-        idata = az.from_netcdf(filepath)
-        model = cls(
-            model_config=json.loads(idata.attrs["model_config"]),
-            sampler_config=json.loads(idata.attrs["sampler_config"]),
-        )
-        model.idata = idata
-        model.build_model()
-        # All previously used data is in idata.
-
-        if model.id != idata.attrs["id"]:
-            raise ValueError(
-                f"The file '{fname}' does not contain an inference data of the same model or configuration as '{cls._model_type}'"
-            )
-
-        return model
 
     @property
     def _serializable_model_config(self):
